Move time-dependent maps into a sub directory
