+++ src/Domain/CoordinateMaps/TimeDependent/Translation.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Domain/CoordinateMaps/Translation.hpp"
+#include "Domain/CoordinateMaps/TimeDependent/Translation.hpp"
 
 #include <ostream>
 #include <pup.h>
