+++ src/Domain/Creators/TimeDependence/UniformTranslation.hpp
@@ -12,7 +12,7 @@
 #include <vector>
 
 #include "Domain/CoordinateMaps/CoordinateMap.hpp"
-#include "Domain/CoordinateMaps/Translation.hpp"
+#include "Domain/CoordinateMaps/TimeDependent/Translation.hpp"
 #include "Domain/Creators/TimeDependence/GenerateCoordinateMap.hpp"
 #include "Domain/Creators/TimeDependence/TimeDependence.hpp"
 #include "Options/Options.hpp"
