+++ tests/Unit/Domain/CoordinateMaps/CMakeLists.txt
@@ -7,20 +7,17 @@ set(LIBRARY_SOURCES
   Test_Affine.cpp
   Test_BulgedCube.cpp
   Test_CoordinateMap.cpp
-  Test_CubicScale.cpp
   Test_DiscreteRotation.cpp
   Test_EquatorialCompression.cpp
   Test_Equiangular.cpp
   Test_Frustum.cpp
   Test_Identity.cpp
   Test_ProductMaps.cpp
-  Test_ProductMapsTimeDep.cpp
   Test_Rotation.cpp
   Test_RotationTimeDep.cpp
   Test_SpecialMobius.cpp
   Test_Tags.cpp
   Test_TimeDependentHelpers.cpp
-  Test_Translation.cpp
   Test_Wedge2D.cpp
   Test_Wedge3D.cpp
   )
@@ -31,3 +28,5 @@ add_test_library(
   "${LIBRARY_SOURCES}"
   "Boost::boost;CoordinateMaps;FunctionsOfTime"
   )
+
+add_subdirectory(TimeDependent)
